+++ modules/perm/types/auth.go
@@ -40,8 +40,8 @@ func (a Auth) Roles() (rs []Role) {
 	if a.Access(RolePowerUserAdmin.Auth()) {
 		rs = append(rs, RolePowerUserAdmin)
 	}
-	if a.Access(RoleLayer2User.Auth()) {
-		rs = append(rs, RoleLayer2User)
+	if a.Access(RoleSideChainUser.Auth()) {
+		rs = append(rs, RoleSideChainUser)
 	}
 
 	return rs