+++ modules/side-chain/keeper/ante.go
@@ -27,7 +27,7 @@ func (dlt ValidateSideChainDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, sim
 				return ctx, err
 			}
 		case *types.MsgTransferSpace:
-			if err := dlt.validateSpaceOwnership(ctx, msg.Sender, msg.SpaceId); err != nil {
+			if err := dlt.validateOnlySpace(ctx, msg.SpaceId); err != nil {
 				return ctx, err
 			}
 			if err := dlt.validateSideChainUserRole(ctx, msg.Sender); err != nil {
@@ -37,7 +37,7 @@ func (dlt ValidateSideChainDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, sim
 				return ctx, err
 			}
 		case *types.MsgCreateBlockHeader:
-			if err := dlt.validateSpaceOwnership(ctx, msg.Sender, msg.SpaceId); err != nil {
+			if err := dlt.validateOnlySpace(ctx, msg.SpaceId); err != nil {
 				return ctx, err
 			}
 			if err := dlt.validateSideChainUserRole(ctx, msg.Sender); err != nil {
@@ -56,22 +56,10 @@ func (dlt ValidateSideChainDecorator) validateOnlySpace(ctx sdk.Context, spaceId
 	return nil
 }
 
-func (dlt ValidateSideChainDecorator) validateSpaceOwnership(ctx sdk.Context, addr string, spaceId uint64) error {
-	accAddr, _ := sdk.AccAddressFromBech32(addr)
-	if !dlt.keeper.HasSpace(ctx, spaceId) {
-		return sdkerrors.Wrapf(types.ErrInvalidSpaceId, "space (%d) does not exist", spaceId)
-	}
-
-	if !dlt.keeper.HasSpaceOfOwner(ctx, accAddr, spaceId) {
-		return sdkerrors.Wrapf(types.ErrInvalidSpaceOwner, "space (%d) is not owned by (%s)", spaceId, addr)
-	}
-	return nil
-}
-
 func (dlt ValidateSideChainDecorator) validateSideChainUserRole(ctx sdk.Context, addr string) error {
 	accAddr, _ := sdk.AccAddressFromBech32(addr)
 	if !dlt.permKeeper.HasSideChainUserRole(ctx, accAddr) {
-		return sdkerrors.Wrapf(types.ErrInvalidSideChainUser, "account (%s) does not have l2 user role", addr)
+		return sdkerrors.Wrapf(types.ErrInvalidSideChainUser, "account (%s) does not have side chain user role", addr)
 	}
 	return nil
 }